requirements/dev-requirements.txt
altgraph==0.17.4
click==8.1.7
colorama==0.4.6
darkdetect==0.7.1
importlib-metadata==7.1.0
imutils==0.5.4
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pefile==2023.2.7
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
pyqtdarktheme==2.1.0
pyqtgraph==0.13.7
pyserial==3.5
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
scipy==1.13.0
six==1.16.0
treelib==1.7.0
zipp==3.18.2
